+++ packages/ui/certd-server/src/plugins/plugin-hipmdnsmgr/access/index.ts
@@ -1,0 +1,1 @@
+export * from './hipmdnsmgr-access.js';
