+++ packages/ui/certd-server/src/plugins/plugin-hipmdnsmgr/dns-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './hipmdnsmgr-dns-provider.js';
