+++ packages/ui/certd-server/src/plugins/plugin-hipmdnsmgr/index.ts
@@ -1,2 +1,2 @@
-export * from './access.js';
-export * from './dns-provider.js';
+export * from './access/index.js';
+export * from './dns-provider/index.js';
